+++ web-vue/src/pages/node/node-layout/other/script-console.vue
@@ -24,6 +24,7 @@
 </template>
 <script>
 import { mapGetters } from "vuex";
+import { getWebSocketUrl } from "@/utils/const";
 
 export default {
   props: {
@@ -54,10 +55,7 @@ export default {
   computed: {
     ...mapGetters(["getLongTermToken"]),
     socketUrl() {
-      const protocol = location.protocol === "https:" ? "wss://" : "ws://";
-      const domain = window.routerBase;
-      const url = (domain + "/node/script_run").replace(new RegExp("//", "gm"), "/");
-      return `${protocol}${location.host}${url}?userId=${this.getLongTermToken}&id=${this.id}&nodeId=${this.nodeId}&type=nodeScript`;
+      return getWebSocketUrl("/socket/node/script_run", `userId=${this.getLongTermToken}&id=${this.id}&nodeId=${this.nodeId}&type=nodeScript`);
     },
   },
   mounted() {