+++ web-vue/src/pages/node/node-layout/system/log.vue
@@ -25,8 +25,9 @@
   </a-layout>
 </template>
 <script>
-import { getLogList, downloadFile, deleteLog } from "../../../../api/system";
+import { getLogList, downloadFile, deleteLog } from "@/api/system";
 import { mapGetters } from "vuex";
+import { getWebSocketUrl } from "@/utils/const";
 export default {
   props: {
     node: {
@@ -52,10 +53,7 @@ export default {
   computed: {
     ...mapGetters(["getLongTermToken"]),
     socketUrl() {
-      const protocol = location.protocol === "https:" ? "wss://" : "ws://";
-      const domain = window.routerBase;
-      const url = (domain + "/tomcat_log").replace(new RegExp("//", "gm"), "/");
-      return `${protocol}${location.host}${url}?userId=${this.getLongTermToken}&tomcatId=${this.tomcatId}&nodeId=${this.node.id}&type=tomcat`;
+      return getWebSocketUrl("/socket/tomcat_log", `userId=${this.getLongTermToken}&tomcatId=${this.tomcatId}&nodeId=${this.node.id}&type=tomcat`);
     },
   },
   watch: {},